[Wends] [added] temporary additional tabs on People tab, renamed Header to NavBar
